requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
pymongo==4.6.0
langchain-openai==0.0.5
langchain-core==0.1.10
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
